fix: refactor interfaces for Audit log
